% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_table.R
\name{extract_followship}
\alias{extract_followship}
\title{Extract Followship Table}
\usage{
extract_followship(.html, ...)
}
\arguments{
\item{.html}{An `HTML` file from each information request.}

\item{...}{Extra arguments to pass.}
}
\description{
`extract_followship` returns a `data.frame` with the values in the HTML.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}

Brendan Knapp, \email{brendan.knapp@nps.edu}
}
