% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_table.R
\name{extract_html_table}
\alias{extract_html_table}
\title{Extract HTML Table}
\usage{
extract_html_table(.html, ...)
}
\arguments{
\item{.html, }{An `HTML` file like the example}
}
\description{
`extract_table` returns a `data.frame` with the values in the HTML.
}
\examples{
library(COREmisc)

le_html <- htmltools::HTML()
out_table <- extract_html_table(le_html)
head(out(table))

}
\seealso{
Uses funcitons from `magrittr` [magrittr::`%>%`],
`dplyr` [dplyr::bind_rows][dplyr::distinct][dplyr::group_split][dplyr::if_else][dplyr::mutate][dplyr::mutate_if][dplyr::select][dplyr::starts_with],
`purr` [purrr::map][purrr::map_dfr],
`rlang` [rlang::`!!!`][rlang::set_names][rlang::syms],
`rvest` [rvest::html_nodes][rvest::html_table],
`tibble` [tibble::as_tibble],
`tidyr` [tidyr::fill][tidyr::gather][tidyr::spread][tidyr::unnest],
`xml2` [xml2::read_html].
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}

Brendan Knapp, \email{brendan.knapp@nps.edu}
}
