% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ips.R
\name{find_ip}
\alias{find_ip}
\title{Find IP address geolocation}
\usage{
find_ip(.df, ...)
}
\arguments{
\item{.df, }{A `data.frame` object with ip address in a column named `Sender_IP`}
}
\description{
`find_ip` returns a `data.frame` with geolocated IP addresses.
}
\examples{
library(COREmisc)
df <- read.csv(file.choose())
out_table <- COREmisc::find_ip(df)

}
\seealso{
Uses funcitons from `magrittr` [magrittr::`%>%`],
`dplyr` [dplyr::left_join][dplyr::distinct]
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
