% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_table.R
\name{get_html}
\alias{get_html}
\title{Get HTML}
\usage{
get_html(file_path)
}
\arguments{
\item{file_path}{the location and name of the file which the data are to be read from.}

\item{...}{Extra arguments to pass.}
}
\description{
Get HTML
}
\keyword{internal}
