% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_table.R
\name{extract_edgelist}
\alias{extract_edgelist}
\title{Extract Edgelist}
\usage{
extract_edgelist(.df, .source, .target, ...)
}
\arguments{
\item{.df}{a \code{data.frame} for parsing.}

\item{.source}{variable name to select as source}

\item{.target}{variable name to select as target}

\item{...}{Extra arguments to pass.}
}
\description{
`extract_edgelist` returns a `data.frame` edgelist from other tables extracted from HTML returns.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
